Rework regex that filters comments/etc from input
I don't entirely understand what Jeff was trying to do with this regex:

    ^\s*[^\#;]\S+

I think he's trying to filter lines that look like comments.  I'm not
sure why he includes the semi, perhaps supporting lisp style comments?
I'm not sure if he was trying to escape the # or if he also wants to
filter lines that start with \.

It also has a bug in that it filters out lines with a single
character, while there aren't any/many single character lines I
tripped over it while testing the regex to see how it behaved.

I've swapped it for one that explicitly does what I find useful:
filtering out empty lines and comment lines that start with two or
more #.  Single # comment lines are kept so that they can be used as
"real-time" documentation while the script is being run.